Copy Pay doc files into Tech Doc Format
Have copied technical documentation files into new tech doc format. The config/tech-doc.yml and source/index-html.md.erb files are changed in line with this. There have been no changes to the content as part of this move. All content changes will be handled in separate PRs in the future.

The gelato documentation will remain in place until this is approved. Once it is approved, the gelato documentation will be replaced by the documentation new format; however the API reference material and API explorer will remain online and in gelato as this functionality is not implemented in the tech doc format yet. Please review.