Support #2788803 Updating documentation
All documentation links are pointing to 1.0.1 which has been deprecated
by v1.0.2.
There is also no documentation specifying the expected values for the
different payment type states when searching for a payment when using
the api.
The payment states can be found when viewing
https://govukpay-docs.cloudapps.digital/#payment-status-lifecycle, but
should also be linked to in the example column for state in the gelato
documentation.
The card type values can be found when viewing
https://docs.payments.service.gov.uk/#card-types, but should also be
linked to in the example column for card_brand in the gelato
documentation.

solo @belindac