Switching to live content overhaul
This is copied from a very large Google Doc which Jon G kindly put in Markdown already (I might commit a few further changes). 

Before I can send this for a content review there are a couple of questions left outstanding so I'm opening this PR to invite people to answer those when they have the time and bandwidth.