Radical changes to 'Versioning'
I'm not necessarily proposing these changes as a "please approve" but rather I hope to use this PR as a place to gather feedback and have a discussion about it. 

I don't believe we should have a 'versioning' section that needs manual updating every time a new version is introduced. For the same reason I don't believe we should have a revision history table (which is in any case empty at the moment).

Making statements like "In the future we will..." is difficult on live documentation, too, if we're not sure of the timeframe. 

I'd also like to know if these points are true:

> We’ll also send you an email to let you know about any new versions we release. 
> We’ll support each version of the early beta GOV.UK Pay API for at least 1 month after we issue an upgrade notice. As the API matures, we may increase this support period to 3 or 6 months.