Remove link to Wikipedia
We shouldn't link to an uncited Wikipedia article. We could write our own content, but it would detract from the point of this page. We could link to somewhere like the FCA's definition, but it's not as good as Wikipedia. On balance, the best/simplest thing to do is to expect that users will look elsewhere for this definition, if they actually need it. If we see a clearly articulated user need for this, we can reintroduce it at a later time.