Small clarifications to the ‘Making payments’ section
Commit 0389a5c750fa170072a973afcc7daa61837cae12 changed the content
of the ‘Making payments’ section. Unfortunately, some information
was removed or put in a place where it doesn’t make as much sense.
Some of the warnings are inaccurate and we’ve been contacted by
confused service users.

• Move `return_url` content back near the beginning of the ‘Making
  payments` section.

• Add back content about the `next_url`.

• Stop warning that the payment ID should be kept secret. It
  doesn’t need to be (we include it in payment page URLs).
  However, keep some content about the `next_url` being sensitive.

• In the ‘Choose the return URL and match users to payments
  section, add a warning that the mechanism used to match returning
  users to payments should be tamper-proof (e.g. by not putting the
  payment ID in an unencrypted cookie), which is possibly what the
  other warning was trying to get at.

This section could do with further attention but these changes make
it a little better for now.