ZD-5266603 Replace mock card 5101110000000004 with 5555555555554444
In the list of mock card numbers, replace 5101110000000004 with
5555555555554444. The BIN range that has 5101110000000004 in it
has now been allocated to a bank.

Both the old and the new card numbers are non-prepaid, non-
corporate Mastercard credit cards.

The sandbox card numbers in connector have already been updated
to recognise the new card number. Card ID did not need updating as
the BIN range that has 5555555555554444 in it was already present
in the test file.